refactor: add dto and module index files
