+++ src/main.ts
@@ -7,22 +7,21 @@ import { resolve } from 'path';
 import * as yaml from 'js-yaml';
 
 import { AppModule } from './app.module';
-import { HttpExceptionFilter } from './filters/http-exception.filter';
+import { HttpExceptionFilter } from './filters';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   const configService = app.get(ConfigService);
-  const swaggerConfig = await readFile(
-    resolve(__dirname, '../doc/api.yaml'),
-    'utf8',
-  );
-  const document = yaml.load(swaggerConfig);
 
+  const openAPI = await readFile(resolve(__dirname, '../doc/api.yaml'), 'utf8');
+  const document = yaml.load(openAPI);
   SwaggerModule.setup('doc', app, document);
+
   app.useGlobalPipes(
     new ValidationPipe({ whitelist: true, forbidNonWhitelisted: true }),
   );
   app.useGlobalFilters(new HttpExceptionFilter());
+
   await app.listen(configService.get('PORT') || 4000);
 }
 bootstrap();
