+++ src/modules/albums/album.module.ts
@@ -1,11 +1,9 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from '../database/database.module';
 import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
 
 @Module({
-  imports: [DatabaseModule],
   controllers: [AlbumController],
   providers: [AlbumService],
   exports: [AlbumService],
