+++ src/modules/albums/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './create-album.dto';
+export * from './update-album.dto';
