+++ src/modules/artists/artist.controller.ts
@@ -12,8 +12,7 @@ import {
 } from '@nestjs/common';
 
 import { ArtistService } from './artist.service';
-import { CreateArtistDto } from './dto/create-artist.dto';
-import { UpdateArtistDto } from './dto/update-artist.dto';
+import { CreateArtistDto, UpdateArtistDto } from './dto';
 import { ArtistEntity } from './entities/artist.entity';
 
 @Controller('artist')
