+++ src/modules/artists/artist.module.ts
@@ -1,11 +1,9 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from '../database/database.module';
 import { ArtistService } from './artist.service';
 import { ArtistController } from './artist.controller';
 
 @Module({
-  imports: [DatabaseModule],
   controllers: [ArtistController],
   providers: [ArtistService],
   exports: [ArtistService],
