+++ src/modules/artists/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './create-artist.dto';
+export * from './update-artist.dto';
