+++ src/modules/favorites/favorite.module.ts
@@ -1,11 +1,9 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from '../database/database.module';
 import { FavoriteService } from './favorite.service';
 import { FavoriteController } from './favorite.controller';
 
 @Module({
-  imports: [DatabaseModule],
   controllers: [FavoriteController],
   providers: [FavoriteService],
 })
