+++ src/modules/prisma/prisma.module.ts
@@ -1,6 +1,7 @@
-import { Module } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
 import { PrismaService } from './prisma.service';
 
+@Global()
 @Module({
   providers: [PrismaService],
   exports: [PrismaService],
