+++ src/modules/tracks/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './create-track.dto';
+export * from './update-track.dto';
