+++ src/modules/tracks/track.controller.ts
@@ -12,8 +12,7 @@ import {
 } from '@nestjs/common';
 
 import { TrackService } from './track.service';
-import { CreateTrackDto } from './dto/create-track.dto';
-import { UpdateTrackDto } from './dto/update-track.dto';
+import { CreateTrackDto, UpdateTrackDto } from './dto';
 import { TrackEntity } from './entities/track.entity';
 
 @Controller('track')
