+++ src/modules/users/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './create-user.dto';
+export * from './update-password.dto';
