+++ src/modules/users/user.module.ts
@@ -1,12 +1,9 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from '../database/database.module';
-import { PrismaModule } from '../prisma/prisma.module';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
 
 @Module({
-  imports: [DatabaseModule, PrismaModule],
   controllers: [UserController],
   providers: [UserService],
 })
